pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "COMP0188-CW2"
version = "0.0.1"
dependencies = [
    "h5py",
    "matplotlib",
    "numpy",
    "pandas",
    "piqa",
    "seaborn",
    "setuptools",
    "torch",
    "torcheval",
    "torchvision",
    "umap-learn",
    "tqdm",
    "wandb",
]
requires-python = ">= 3.10" # 3.9 for PEP 585 and 3.10 itertools.pairwise haha
authors = [
  {name = "Iyassou Shimels", email = "s.iyassou@gmail.com"},
]
description = "COMP0188 Deep Representations and Learning – Coursework 2"
readme = "README.md"
license = {file = "LICENSE.txt"}
keywords = ["VAE", "Behaviour Cloning", "Robotics", "Computer Vision"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]

[project.optional-dependencies]
dev = [
  "pytest"
]

[project.urls]
Repository = "https://github.com/iyassou/COMP0188-CW2.git"